+++ src/app/src/components/ClaimFacility.jsx
@@ -15,8 +15,8 @@ import ClaimFacilityStepper from '../components/ClaimFacilityStepper';
 import {
     fetchClaimFacilityData,
     clearClaimFacilityDataAndForm,
-    fetchParentCompanyOptions,
 } from '../actions/claimFacility';
+import { fetchParentCompanyOptions } from '../actions/filterOptions';
 
 import { facilityDetailsPropType } from '../util/propTypes';
 
@@ -124,7 +124,9 @@ ClaimFacility.propTypes = {
 function mapStateToProps({
     claimFacility: {
         facilityData: { data, fetching, error },
-        parentCompanyOptions: { fetching: fetchingParentCompanyOptions },
+    },
+    filterOptions: {
+        parentCompanies: { fetching: fetchingParentCompanyOptions },
     },
     auth: {
         user: { user },