+++ src/app/src/reducers/FilterOptionsReducer.js
@@ -17,6 +17,9 @@ import {
     startFetchSectorOptions,
     failFetchSectorOptions,
     completeFetchSectorOptions,
+    startFetchParentCompanyOptions,
+    failFetchParentCompanyOptions,
+    completeFetchParentCompanyOptions,
     startFetchFacilityProcessingTypeOptions,
     failFetchFacilityProcessingTypeOptions,
     completeFetchFacilityProcessingTypeOptions,
@@ -28,42 +31,47 @@ import {
 
 const initialState = Object.freeze({
     contributors: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
     lists: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
     contributorTypes: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
     countries: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
     sectors: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
+        fetching: false,
+        error: null,
+    }),
+    parentCompanies: Object.freeze({
+        data: null,
         fetching: false,
         error: null,
     }),
     facilityProcessingType: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
     productType: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
     numberOfWorkers: Object.freeze({
-        data: Object.freeze([]),
+        data: null,
         fetching: false,
         error: null,
     }),
@@ -179,6 +187,31 @@ export default createReducer(
                     fetching: { $set: false },
                     error: { $set: null },
                     data: { $set: payload },
+                },
+            }),
+        [startFetchParentCompanyOptions]: state =>
+            update(state, {
+                parentCompanies: {
+                    fetching: { $set: true },
+                    error: { $set: initialState.parentCompanies.error },
+                },
+            }),
+        [failFetchParentCompanyOptions]: (state, error) =>
+            update(state, {
+                parentCompanies: {
+                    fetching: {
+                        $set: initialState.parentCompanies.fetching,
+                    },
+                    error: { $set: error },
+                },
+            }),
+        [completeFetchParentCompanyOptions]: (state, data) =>
+            update(state, {
+                parentCompanies: {
+                    data: { $set: data },
+                    fetching: {
+                        $set: initialState.parentCompanies.fetching,
+                    },
                 },
             }),
         [startFetchFacilityProcessingTypeOptions]: state =>